requirements-prod.txt
# Production Dependencies for SEC Latent Application
# Security-hardened production-ready packages

# Web Server & ASGI
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# Monitoring & Observability
sentry-sdk[fastapi,celery,redis,sqlalchemy]==1.39.1
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-celery==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1

# Health Checks & Metrics
py-healthcheck==1.10.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9

# Caching
redis[hiredis]==5.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Task Queue
celery[redis]==5.3.4
flower==2.0.1

# Logging
python-json-logger==2.0.7
structlog==23.2.0

# Performance
orjson==3.9.10  # Faster JSON serialization
httpx[http2]==0.25.2  # HTTP/2 support

# Backup & Recovery
boto3==1.34.3  # AWS S3 for backups
